+++ routers/users.js
@@ -1,12 +1,12 @@
 const express = require('express');
 const { signupController, loginController, logoutController } = require('../controllers/userController');
 const tryCatchWrapper = require('../utils/tryCatchWrapper');
-const { loginValidation } = require('../middlewares/validationLoginMiddlewate');
+const { loginValidation, signupValidation } = require('../middlewares/validationMiddleware');
 const { authMiddleware } = require('../middlewares/authMiddleware');
 
 const router = express.Router();
 
-router.post('/signup', loginValidation, tryCatchWrapper(signupController));
+router.post('/signup', signupValidation, tryCatchWrapper(signupController));
 
 router.post('/login', loginValidation, tryCatchWrapper(loginController));
